// server/src/models/Board/BoardKey.ts
export enum BoardKey {
  id = "id",
  title = "title",
  theme = "theme",
  idUser = "idUser",
  idLists = "idLists",
  idLabels = "idLabels",
  isStarred = "isStarred",
  dateCreation = "dateCreation",
  dateLastView = "dateLastView",
  dateLastActivity = "dateLastActivity",
}


export default BoardKey